Fixed possible buffer overflow on many strncat()

+++ src/common/grayproto.h
@@ -59,8 +59,8 @@ struct NDWORD
 
 #define	NCHAR	NWORD			// a UNICODE text char on the network.
 
-extern int CvtSystemToNUNICODE( NCHAR * pOut, int iSizeOutChars, LPCTSTR pInp, int iSizeInBytes );
-extern int CvtNUNICODEToSystem( TCHAR * pOut, int iSizeOutBytes, const NCHAR * pInp, int iSizeInChars );
+extern size_t CvtSystemToNUNICODE(NCHAR *pOut, int iSizeOutChars, LPCTSTR pInp, int iSizeInBytes);
+extern size_t CvtNUNICODEToSystem(TCHAR *pOut, int iSizeOutBytes, const NCHAR *pInp, int iSizeInChars);
 
 class CLanguageID
 {
